+++ pidx/data_handle/PIDX_idx_data_structs.h
@@ -223,8 +223,8 @@ struct idx_file_struct
 
   int bits_per_block;
   int blocks_per_file;
-  unsigned long long bounds[PIDX_MAX_DIMENSIONS];
-  unsigned long long box_bounds[PIDX_MAX_DIMENSIONS];
+  size_t bounds[PIDX_MAX_DIMENSIONS];
+  size_t box_bounds[PIDX_MAX_DIMENSIONS];
   double transform[16];
   char bitSequence[512];
   char bitPattern[512];
@@ -239,7 +239,7 @@ struct idx_file_struct
   int compression_type;
   int compression_factor;
   float compression_bit_rate;
-  unsigned long long chunk_size[PIDX_MAX_DIMENSIONS];
+  size_t chunk_size[PIDX_MAX_DIMENSIONS];
 
   int file_zero_merge;
