+++ pidx/data_handle/PIDX_memory_layout_data_structs.h
@@ -37,8 +37,8 @@ struct PIDX_Ndim_empty_patch_struct
 {
   int rank;
   int is_boundary_patch;
-  unsigned long long offset[PIDX_MAX_DIMENSIONS];       ///< offset of the data chunk (of PIDX_MAX_DIMENSIONS dimension)
-  unsigned long long size[PIDX_MAX_DIMENSIONS];         ///< size (extents) in each of the dimensions for the data chunk
+  off_t offset[PIDX_MAX_DIMENSIONS];       ///< offset of the data chunk (of PIDX_MAX_DIMENSIONS dimension)
+  size_t size[PIDX_MAX_DIMENSIONS];         ///< size (extents) in each of the dimensions for the data chunk
 };
 typedef struct PIDX_Ndim_empty_patch_struct* Ndim_empty_patch;
 
@@ -47,7 +47,7 @@ typedef struct PIDX_Ndim_empty_patch_struct* Ndim_empty_patch;
 /// Struct to store the restructured grid
 struct PIDX_grid_struct
 {
-  unsigned long long patch_size[PIDX_MAX_DIMENSIONS];
+  size_t patch_size[PIDX_MAX_DIMENSIONS];
   int total_patch_count[PIDX_MAX_DIMENSIONS];
   Ndim_empty_patch* patch;
 };
@@ -58,8 +58,8 @@ typedef struct PIDX_grid_struct* PIDX_restructured_grid;
 struct PIDX_patch_struct
 {
   int particle_count;
-  unsigned long long offset[PIDX_MAX_DIMENSIONS];       ///< offset of the data chunk (of PIDX_MAX_DIMENSIONS dimension) in the 3D global space
-  unsigned long long size[PIDX_MAX_DIMENSIONS];         ///< size (extents) in each of the dimensions for the data chunk
+  off_t offset[PIDX_MAX_DIMENSIONS];       ///< offset of the data chunk (of PIDX_MAX_DIMENSIONS dimension) in the 3D global space
+  size_t size[PIDX_MAX_DIMENSIONS];         ///< size (extents) in each of the dimensions for the data chunk
   unsigned char* buffer;                                ///< the data buffer
 };
 typedef struct PIDX_patch_struct* PIDX_patch;
